fix SpotlightProvider Loading State(#63)
